broken DXTextureAdd

+++ tomb3decomp/specific/picture.cpp
@@ -234,7 +234,7 @@ void ConvertSurfaceToTextures16Bit(LPDIRECTDRAWSURFACEX surf)
 	memcpy(&desc2, &desc, sizeof(DDSURFACEDESC));
 	MemBlt(dest, 0, 0, 256, 256, 256, source, 0, 0, desc2);
 	surf->Unlock(0);
-	pIndices[0] = DXTextureAdd(256, 256, (ushort*)dest, PictureTextures, bitcnt, 16);
+	pIndices[0] = DXTextureAdd(256, 256, (uchar*)dest, PictureTextures, bitcnt, 16);
 
 	memset(&desc, 0, sizeof(DDSURFACEDESCX));
 	desc.dwSize = sizeof(DDSURFACEDESCX);
@@ -243,7 +243,7 @@ void ConvertSurfaceToTextures16Bit(LPDIRECTDRAWSURFACEX surf)
 	source = (char*)desc.lpSurface;
 	MemBlt(dest, 0, 0, 256, 256, 256, source, 256, 0, desc2);
 	surf->Unlock(0);
-	pIndices[1] = DXTextureAdd(256, 256, (ushort*)dest, PictureTextures, bitcnt, 16);
+	pIndices[1] = DXTextureAdd(256, 256, (uchar*)dest, PictureTextures, bitcnt, 16);
 
 	memset(&desc, 0, sizeof(DDSURFACEDESCX));
 	desc.dwSize = sizeof(DDSURFACEDESCX);
@@ -254,7 +254,7 @@ void ConvertSurfaceToTextures16Bit(LPDIRECTDRAWSURFACEX surf)
 	memcpy(&desc2, &desc, sizeof(DDSURFACEDESCX));
 	MemBlt(dest, 128, 0, 128, 224, 256, source, 512, 256, desc2);
 	surf->Unlock(0);
-	pIndices[2] = DXTextureAdd(256, 256, (ushort*)dest, PictureTextures, bitcnt, 16);
+	pIndices[2] = DXTextureAdd(256, 256, (uchar*)dest, PictureTextures, bitcnt, 16);
 
 	memset(&desc, 0, sizeof(DDSURFACEDESCX));
 	desc.dwSize = sizeof(DDSURFACEDESCX);
@@ -263,7 +263,7 @@ void ConvertSurfaceToTextures16Bit(LPDIRECTDRAWSURFACEX surf)
 	source = (char*)desc.lpSurface;
 	MemBlt(dest, 0, 0, 256, 224, 256, source, 0, 256, desc2);
 	surf->Unlock(0);
-	pIndices[3] = DXTextureAdd(256, 256, (ushort*)dest, PictureTextures, bitcnt, 16);
+	pIndices[3] = DXTextureAdd(256, 256, (uchar*)dest, PictureTextures, bitcnt, 16);
 
 	memset(&desc, 0, sizeof(DDSURFACEDESCX));
 	desc.dwSize = sizeof(DDSURFACEDESCX);
@@ -272,7 +272,7 @@ void ConvertSurfaceToTextures16Bit(LPDIRECTDRAWSURFACEX surf)
 	source = (char*)desc.lpSurface;
 	MemBlt(dest, 0, 0, 256, 224, 256, source, 256, 256, desc2);
 	surf->Unlock(0);
-	pIndices[4] = DXTextureAdd(256, 256, (ushort*)dest, PictureTextures, bitcnt, 16);
+	pIndices[4] = DXTextureAdd(256, 256, (uchar*)dest, PictureTextures, bitcnt, 16);
 
 	free(dest);
 }
